add make_package script
The rationale for this script is that using a make target (like
'make package') is really stretching the point as there are not really
dependencies to check.

So we might as well do it in a shell script instead of abusing a Makefile.

Also, this makes it clear that there are none of the usual devel tools
needed to prepare a tar archive.