rework log creation to work better with merge commits
This reworks the way merge commit messages are handled. It basically
collects all commit messages belonging to a merge (even those directly in
the merge commit itself) and groups them into a common changelog entry.

The date and author of the merge commit will be used for the changelog.

As this will collect the same log message twice in some cases, duplicate
lines are removed.